refactor : 랜덤변수에 따라 움직이는 move를 car의 책임으로 분리
